requirements.txt
visdom
torchvision
scikit-image
tqdm
fire
opencv-python==4.5.5.64
matplotlib
torchnet==0.0.4
ipdb
cython